requirements.txt
flask==2.3.3
flask-pymongo==2.3.0
pymongo[srv]==4.6.1
flask-marshmallow==0.15.0
marshmallow==3.20.2
flask-swagger-ui==4.11.1
python-dotenv==1.0.0
werkzeug==2.3.7 